# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Todo Task Manager API",
    "description": "Full-stack todo task management application with user authentication and JWT-based authorization. Built with FastAPI, SQLModel, and Neon PostgreSQL.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@todoapp.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.todoapp.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/auth/signup": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User signup",
        "description": "Register a new user account. Password must be at least 8 characters with uppercase, lowercase, and number.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              },
              "example": {
                "email": "user@example.com",
                "password": "SecurePass123",
                "name": "John Doe"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (weak password, invalid email format)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "weak_password": {
                    "value": {
                      "detail": "Password must contain at least one uppercase letter",
                      "status": 400
                    }
                  },
                  "invalid_email": {
                    "value": {
                      "detail": "Invalid email format",
                      "status": 400
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "Email already registered",
                  "status": 409
                }
              }
            }
          }
        }
      }
    },
    "/auth/signin": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User signin",
        "description": "Authenticate existing user and receive JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SigninRequest"
              },
              "example": {
                "email": "user@example.com",
                "password": "SecurePass123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signin successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "Invalid email or password",
                  "status": 401
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/signup or /auth/signin"
      }
    },
    "schemas": {
      "SignupRequest": {
        "type": "object",
        "required": ["email", "password", "name"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 3,
            "maxLength": 255,
            "description": "Valid email address"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "description": "Password (min 8 chars, must include uppercase, lowercase, and number)"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "User's display name"
          }
        }
      },
      "SigninRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Registered email address"
          },
          "password": {
            "type": "string",
            "description": "User password"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "token": {
            "type": "string",
            "description": "JWT token (valid for 7 days)"
          },
          "expires_in": {
            "type": "integer",
            "description": "Token expiration in seconds",
            "example": 604800
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "description": "User email"
          },
          "name": {
            "type": "string",
            "description": "User display name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation timestamp"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          }
        }
      }
    }
  }
}